requirements.txt
flask==3.0.3
pytest==8.2.0
pytest-flask==1.3.0
requests==2.32.3    
google-genai